pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "The second project which s aim to show my skills in coding, although to get a practice."
authors = ["bogdan <telioko@tutanota.com>"]
license = "GNU"
readme = "README.md"
packages = [{include = "gendiff"}]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
pip = "^23.0.1"
pytest-cov = "^4.0.0"
pyyaml = "^6.0"

[tool.poetry.scripts]
gendiff = 'gendiff.cli:parse'

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
